Write triggers function merged. 100Lines to successfull. Run_Twice
Write and update functions merged and now only write_triggers exists, with a flag to differentiate between running and finished. Replaced the 50Lines used to verify a mohid run, by 100 lines. Completed Run_Twice feature. if active, the model searches for the trigger on the second time it runs a specific day.